# docs/swagger.yaml
basePath: /
definitions:
...
  models.Bahan:
    properties:
      batas_minimum:
        type: number
      harga_saat_ini:
        description: Update otomatis dari pembelian terakhir
        type: number
      id:
        type: integer
      nama_bahan:
        type: string
      satuan:
        description: gr, ml, pcs
        type: string
      stok:
        type: number
      urutan:
        type: integer
    type: object
...
  models.Barang:
    properties:
      Urutan:
        type: integer
      hargaDefault:
        format: float64
        type: number
      id:
        type: integer
      kebutuhan_adonan:
        description: Berapa gram / fraksi per 1 roti
        type: number
      kemasan_detail:
        items:
          $ref: '#/definitions/models.BarangKemasan'
        type: array
      masa_simpan:
        description: Default 2 hari
        type: integer
      metode_konversi:
        description: '"Gram" atau "Pcs"'
        type: string
      namaBarang:
        type: string
      resep:
        $ref: '#/definitions/models.Resep'
      resep_id:
        description: TAMBAHAN MODUL INVENTORY
        type: integer
    type: object
...
  models.BarangKemasan:
    properties:
      bahan:
        $ref: '#/definitions/models.Bahan'
      bahan_id:
        type: integer
      barang_id:
        type: integer
      id:
        type: integer
      kebutuhan:
        description: (Bisa pecahan, misal 0.25)
        type: number
    type: object
...
  models.MessageResponse:
    properties:
      message:
        example: Operasi berhasil dilakukan
        type: string
    type: object
  models.NextNotaResponse:
    properties:
      no_nota:
        example: NT/20260427/15-0017
        type: string
    type: object
  models.Nota:
    properties:
      IsHarianSnapshot:
        type: boolean
      NamaTokoSnapshot:
        description: SNAPSHOT UNTUK MENGUNCI SEJARAH
        type: string
      SiklusSnapshot:
        type: string
      assigned_to:
        type: integer
      created_at:
        type: string
      created_by:
        description: PELACAK SALES
        type: integer
      details:
        items:
          $ref: '#/definitions/models.NotaDetail'
        type: array
      id:
        type: integer
      is_lunas:
        description: '''KIRIM'' atau ''SELESAI'''
        type: boolean
      jumlahKirim:
        description: Hasil Perhitungan
        type: number
      jumlahRetur:
        description: Total harga retur (Semua barang)
        type: number
      noNota:
        type: string
      status:
        type: string
      tanggalKirim:
        type: string
      toko:
        $ref: '#/definitions/models.Toko'
      tokoID:
        type: integer
      total_diskon:
        description: <--- BARU
        type: number
      total_voucher:
        description: <--- BARU
        type: number
      totalBayar:
        description: JumlahKirim - JumlahRetur
        type: number
      updated_at:
        type: string
    type: object
  models.NotaDetail:
    properties:
      NamaBarangSnapshot:
        description: SNAPSHOT UNTUK MENGUNCI SEJARAH
        type: string
      banyakKirim:
        type: integer
      banyakRetur:
        type: integer
      barang:
        $ref: '#/definitions/models.Barang'
      barangID:
        type: integer
      hargaJual:
        type: number
      hargaKirim:
        description: BanyakKirim * HargaJual
        type: number
      hargaRetur:
        description: BanyakRetur * HargaJual
        type: number
      id:
        type: integer
      notaID:
        type: integer
    type: object
  models.NotaDetailInput:
    properties:
      banyak_kirim:
        description: <-- Disesuaikan
        example: 100
        type: integer
      banyak_retur:
        description: <-- Disesuaikan
        example: 5
        type: integer
      barang_id:
        example: 12
        type: integer
      harga_jual:
        description: <-- Disesuaikan
        example: 15000
        type: number
      id:
        description: 0 saat Create baru, isi ID asli saat Update
        example: 0
        type: integer
    required:
    - banyak_kirim
    - barang_id
    - harga_jual
    type: object
  models.NotaInput:
    properties:
      assigned_to:
        example: 2
        type: integer
      details:
        items:
          $ref: '#/definitions/models.NotaDetailInput'
        type: array
      is_lunas:
        example: false
        type: boolean
      no_nota:
        example: NT/20260427/15-0017
        type: string
      status:
        enum:
        - KIRIM
        - DIBATALKAN
        - SELESAI
        example: KIRIM
        type: string
      tanggal_kirim:
        example: "2026-04-27"
        type: string
      toko_id:
        example: 5
        type: integer
      total_diskon:
        example: 0
        type: number
      total_voucher:
        example: 50000
        type: number
    required:
    - details
    - no_nota
    - status
    - tanggal_kirim
    - toko_id
    type: object
...
  models.Resep:
    properties:
      bahan_detail:
        items:
          $ref: '#/definitions/models.ResepBahan'
        type: array
      id:
        type: integer
      nama_resep:
        type: string
      target_gramasi:
        description: Total adonan matang dr 1 Resep
        type: number
    type: object
  models.ResepBahan:
    properties:
      bahan:
        $ref: '#/definitions/models.Bahan'
      bahan_id:
        type: integer
      id:
        type: integer
      kebutuhan:
        description: Butuh berapa gr/ml/pcs
        type: number
      resep_id:
        type: integer
    type: object
...
  models.Toko:
    properties:
      IsHarian:
        type: boolean
      alamat:
        type: string
      id:
        type: integer
      namaToko:
        type: string
      noTelp:
        type: string
      siklusDua:
        type: boolean
      siklusJumatSelasa:
        type: boolean
      siklusKamisSenin:
        description: Flag Siklus
        type: boolean
      siklusSabtuRabu:
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact:
    name: Aristo Budiman
  description: Dokumentasi API Internal untuk Sistem Inventory, Nota, dan Kas.
  title: Tiara Bakery Master API
  version: "1.0"
paths:
...
  /notas:
    get:
      consumes:
      - application/json
      description: Menampilkan daftar keseluruhan riwayat nota kiriman dari yang terbaru
        (descending).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Nota'
            type: array
        "500":
          description: Kesalahan server
          schema:
            additionalProperties: true
            type: object
      summary: Ambil Semua Riwayat Nota
      tags:
      - 08. Nota Reguler
    post:
      consumes:
      - application/json
      description: Memasukkan data nota kiriman harian beserta detail array barangnya
        ke database. Akan memicu trigger sinkronisasi brankas (Kas) jika nota ditandai
        lunas.
      parameters:
      - description: Data lengkap nota dan detail barang
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.NotaInput'
      produces:
      - application/json
      responses:
        "200":
          description: 'Pesan: Nota berhasil disimpan!'
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Format data JSON tidak valid
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Kesalahan eksekusi database / kas
          schema:
            additionalProperties: true
            type: object
      summary: Buat Nota Reguler baru
      tags:
      - 08. Nota Reguler
  /notas/{id}:
    get:
      consumes:
      - application/json
      description: Menampilkan data spesifik sebuah nota beserta hierarki relasinya
        (Detail Nota, Barang terkait, dan Toko).
      parameters:
      - description: ID Nota
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Nota'
        "404":
          description: Nota tidak ditemukan
          schema:
            additionalProperties: true
            type: object
      summary: Ambil Detail Nota (Berdasarkan ID)
      tags:
      - 08. Nota Reguler
    put:
      consumes:
      - application/json
      description: Mengubah data nota kiriman historis berdasarkan ID. Endpoint ini
        sangat krusial karena otomatis menghitung selisih uang jika status lunas/tidak
        lunas berubah.
      parameters:
      - description: ID Nota
        in: path
        name: id
        required: true
        type: integer
      - description: Data nota yang akan direvisi
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.NotaInput'
      produces:
      - application/json
      responses:
        "200":
          description: 'Pesan: Nota berhasil diupdate!'
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Format data JSON tidak valid
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Nota tidak ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Kesalahan eksekusi database / kas
          schema:
            additionalProperties: true
            type: object
      summary: Ubah data Nota Reguler
      tags:
      - 08. Nota Reguler
  /notas/{id}/batal:
    put:
      consumes:
      - application/json
      description: Mengubah status nota menjadi 'DIBATALKAN'. Jika sebelumnya lunas,
        sistem otomatis akan menarik uang dari brankas/kas untuk menjaga integritas
        pembukuan.
      parameters:
      - description: ID Nota
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Pesan: Nota berhasil dibatalkan'
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Nota tidak ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Kesalahan saat membatalkan transaksi kas
          schema:
            additionalProperties: true
            type: object
      summary: Batalkan Nota Reguler (Rollback)
      tags:
      - 08. Nota Reguler
  /notas/{id}/pulihkan:
    put:
      consumes:
      - application/json
      description: Mengembalikan status nota dari 'DIBATALKAN' menjadi 'KIRIM'. Menginjeksi
        ulang uang ke brankas jika nota tercatat lunas.
      parameters:
      - description: ID Nota
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'Pesan: Nota berhasil dipulihkan'
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "404":
          description: Nota tidak ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Kesalahan server
          schema:
            additionalProperties: true
            type: object
      summary: Pulihkan Nota Reguler
      tags:
      - 08. Nota Reguler
  /notas/next-number:
    get:
      consumes:
      - application/json
      description: 'Meng-generate nomor nota urut yang pintar berdasarkan tanggal
        dan ID toko (Cth: NT/20260427/15-0017)'
      parameters:
      - description: Filter berdasarkan ID Toko (Jika 0 = Pabrik)
        in: query
        name: toko_id
        type: string
      - description: 'Format: YYYY-MM-DD (Default: Hari ini)'
        in: query
        name: tanggal
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NextNotaResponse'
      summary: Dapatkan nomor nota berikutnya
      tags:
      - 08. Nota Reguler
...
  /sales/kunjungan/{toko_id}:
    get:
      consumes:
      - application/json
      description: Mengecek apakah sebuah toko memiliki nota berstatus 'KIRIM' namun
        belum diselesaikan proses returnya. Digunakan untuk memblokir pembuatan nota
        baru oleh sales.
      parameters:
      - description: ID Toko
        in: path
        name: toko_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Nota'
            type: array
      summary: Sidak Dosa Retur Toko
      tags:
      - 16. Distribusi Lapangan
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: "Bearer {token_jwt}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"